Extract cursor movement helpers to a module
