+++ xray_core/src/lib.rs
@@ -2,6 +2,7 @@ extern crate rand;
 extern crate futures;
 
 mod notify_cell;
+mod movement;
 mod tree;
 pub mod buffer;
 pub mod editor;
